requirements.txt
requests~=2.28.2
websocket-client~=1.5.1
protobuf==3.6.0
loguru~=0.7.0
pydantic~=1.10.6
PyExecJS~=1.5.1